+++ CHANGES.md
@@ -12,6 +12,7 @@ v0.2.2 (in progress)
   - `max`
   - `sum`
   - `find`
+  - `indexOf`
   - `toArray`
 - added `Lazy(object).watch` to monitor changes to a property as a sequence (see #49)
 - added `Sequence#chunk`, `Sequence#tap`, and `Sequence#ofType`