+++ CHANGES.md
@@ -13,6 +13,7 @@ v0.2.2 (in progress)
   - `sum`
   - `find`
   - `indexOf`
+  - `contains`
   - `toArray`
 - added `Lazy(object).watch` to monitor changes to a property as a sequence (see #49)
 - added `Sequence#chunk`, `Sequence#tap`, and `Sequence#ofType`