+++ CHANGES.md
@@ -16,6 +16,7 @@ v0.2.2 (in progress)
   - `contains`
   - `toArray`
   - `toObject`
+  - `join`/`toString`
 - added `Lazy(object).watch` to monitor changes to a property as a sequence (see #49)
 - added `Sequence#chunk`, `Sequence#tap`, and `Sequence#ofType`
 - deprecated `Lazy.events` in favor of `NodeSequence#on` (note: I should probably rename `NodeSequence` to `DOMSequence`)